feat: added some dialogs and UI design elements

+++ lib/ui/views/home/task_list/task_list_view_model.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/cupertino.dart';
+import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:provider/provider.dart';
 import 'package:withu_todo/non_ui/constants/strings.dart';
 import 'package:withu_todo/non_ui/models/task/task.dart';
+import 'package:withu_todo/non_ui/services/dialog/dialog_service.dart';
 import 'package:withu_todo/non_ui/services/navigation/navigation_service.dart';
 import 'package:withu_todo/non_ui/services/task/task_service.dart';
 import 'package:withu_todo/non_ui/states/busy_state.dart';
@@ -9,6 +11,7 @@ import 'package:withu_todo/ui/views/home/task_form/task_form_view.dart';
 
 class TaskListViewModel extends BusyModel {
   final NavigationService _navigationService = NavigationService.instance;
+  final DialogService _dialogService = DialogService.instance;
 
   int _index;
 
@@ -27,10 +30,19 @@ class TaskListViewModel extends BusyModel {
       ? AppStrings.addYourFirstTask
       : AppStrings.noCompletedTasksFound;
 
-  void delete(Task task, BuildContext context) {
-    final TaskService _taskService =
-        Provider.of<TaskService>(context, listen: false);
-    _taskService.deleteTask(task);
+  void delete(Task task, BuildContext context) async {
+    var dialogResult = await _dialogService.showDialog(
+      title: AppStrings.warningTitle,
+      description: '${task.title} will be deleted.',
+      buttonTitle: 'CONFIRM',
+    );
+    if (dialogResult.confirmed == true) {
+      final TaskService _taskService =
+          Provider.of<TaskService>(context, listen: false);
+      _taskService.deleteTask(task);
+    } else {
+      debugPrint('User cancelled the dialog');
+    }
   }
 
   void toggleComplete(Task task, BuildContext context) {
@@ -38,6 +50,11 @@ class TaskListViewModel extends BusyModel {
     final TaskService _taskService =
         Provider.of<TaskService>(context, listen: false);
     _taskService.updateTask(task);
+    EasyLoading.showToast(
+        task.isCompleted
+            ? AppStrings.toastComplete
+            : AppStrings.toastIncomplete,
+        toastPosition: EasyLoadingToastPosition.bottom);
   }
 
   void navigateToTaskForm({Task task}) {
